+++ zikzak_inappwebview_platform_interface/lib/src/webview_environment/main.dart
@@ -1,2 +1,3 @@
 export 'platform_webview_environment.dart';
-export 'webview_environment_settings.dart' show WebViewEnvironmentSettings;
+export 'webview_environment_settings.dart'
+    show HostResourceAccessKind, VirtualHostMapping, WebViewEnvironmentSettings;